test: narrow hotspot boundaries
